Create README.md
This is an app we made during week 3 of the FNB app academy. It is a tour guide for anyone interested in visiting the fictional pine city zoo. It has functionalities like checking the weather, viewing the map and submitting feedback. These are purely front-end so there are limitations : there is no connected backend. It was developed using HTML5 and CSS.